refactor: split parquet reader nested stream helpers
